Update DoubleMetaphone.sas
Correct error in processing "LL".  If Spanish, DMPV1 is the anglicized version, pronounced as "L", and DMPV2 is the Spanish version, "LL" treated as "Y", and not encoded as "Y" is treated as a vowel.  This error is present in many open source documents, but Open Refine's version of Metaphone 3, (on GitHub) shows the correct handling (along with other special cases not included here, for consistency with other Double Metaphone implementations).